src/requirements.txt
aio-pika==9.5.5
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aioredis==1.3.1
aiormq==6.8.1
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
better-profanity==0.7.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dnspython==2.7.0
dotenv==0.9.9
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.12
fastapi-cache==0.1.0
fastapi-cache2==0.2.1
fastapi-utils==0.8.0
frozenlist==1.5.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
hiredis==3.1.0
idna==3.10
PyJWT==2.9.0
kafka-python==2.1.5
libmagic==1.0
magic-filter==1.0.9
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.4.3
mypy-extensions==1.0.0
packaging==24.2
pamqp==3.3.0
passlib==1.7.4
pendulum==3.0.0
propcache==0.3.1
psutil==5.9.8
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-magic==0.4.27
python-multipart==0.0.20
redis==5.2.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
time-machine==2.16.0
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
user-agents==2.2.0
uvicorn==0.34.1
yarl==1.19.0
